reafactor: DB 관련 전반적인 코드 수정
엔티티 이름 변경
DailyHydrationRecord -> HydratedRecord

각 테이블의 Column 에 @ColumnInfo 를 추가, name 값에 스네이크 케이스를 적용
엔티티 실제 요소는 카멜케이스 적용

HydratedRecord 테이블과 Beverage 테이블 1:1 관계 구축

+++ app/src/main/java/com/example/android/waterwise/data/BeverageRepository.kt
@@ -8,5 +8,5 @@ interface BeverageRepository {
     suspend fun deleteBeverage(beverage: Beverage)
     fun getAllBeverageStream(): Flow<List<Beverage>>
     fun getAllBeverageWithHydratePresetsStream(): Flow<List<BeverageWithHydratePresets>>
-    fun getBeverageWithHydratePresetsById(id: Int): Flow<BeverageWithHydratePresets>
+    fun getBeverageWithHydratePresetsById(id: Long): Flow<BeverageWithHydratePresets>
 }
